refactor PIAForm Types and interfaces
