+++ src/frontend/src/pages/PIAForm/helpers/enums.ts
@@ -1,0 +1,4 @@
+export enum PiaFormSubmissionTypeEnum {
+  PI,
+  DELEGATED,
+}
